Implementing Transformers for section metadata
